Disable scaling of tablet coordinates towards screen centre.
IOHIDEventDriver uses default value of 15 for AbsoluteAxisBoundsRemovalPercentage. Set it to 0 in our info.plist. This setting configures the amount by which the “live” area of the tablet is shrunk. In practice, this makes the edges of the screen “resist” the mouse cursor from “escaping”, which makes it easier to hit menu bar and dock items quickly and precisely, but does introduce a discontinuity in mouse movement. Setting it to 0 disables this feature/bug.

Bumped kext version number to 1.1b1. Still undecided if I want to make the scaling user-configurable.